hardware_monitor_backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0
psutil==5.9.5
scikit-learn==1.3.0
pandas==2.1.0
numpy==1.25.2
joblib==1.3.2
pythonnet>=3.0.0